+++ src/pages/join/index.tsx
@@ -1,12 +1,12 @@
 import ResponsiveProvider from '../ui/ResponsiveProvider';
 import FormHeader from '../../shared/ui/FormHeader';
-import JoinForm from '../../entities/join/JoinForm';
+import { User } from '../../features';
 
 export default function Join() {
   return (
     <ResponsiveProvider direction={'col'}>
       <FormHeader text={'회원가입'} />
-      <JoinForm />
+      <User.UI.JoinForm />
     </ResponsiveProvider>
   );
 }
