+++ src/pages/login/index.tsx
@@ -1,6 +1,6 @@
 import ResponsiveProvider from '../ui/ResponsiveProvider';
 import FormHeader from '../../shared/ui/FormHeader';
-import LoginForm from '../../entities/login/LoginForm';
+import { LoginForm } from '../../features/user/ui';
 export default function Login() {
   return (
     <ResponsiveProvider direction={'col'}>
